Hi, here's a new version of my python interface structure refinement tool.
I tried to create a somewhat more grown up version. So I removed all the
experimental stuff.
I removed the simulated annealing and the genetic algorithm and just left
the simplex minimization routine, which in my opinion works best.
Most important change is that I finally included parameter statistics /
correlation calculations and sensitivity analyses and plots and so on.
So now the code can be used for serious scientific work ;-) (statistics
are based on the math as explained in UCODE (USGS) manual, but I kept is
a bit simpler)
Cheers, Frank (FHe)